refactor: extract BaseGitHubProvider to reduce duplication [swarm]
